Beginnings of a deferred renderer: a tiny gbuffer.
Still very not-finished... While it does seem to work correctly, it is still unstable. Resizing the window causes the validation layers to crash now, which is kind of weird, as it worked fine before, and none of the swapchain (re)-creation code has been changed...

Not fixing it today, but is the priority TODO for now.

The DEFERRED define switches between the old and new render path. Once this new path is stable, the old will be deleted and the bozo and bz namespaces can be reunified.

The next steps after that are to expand the gbuffer, implement proper ubo's to send fx. simple light data to the 'deferred' pass.

And then when I get old and tired I might get to actually implement some rendering techniques!