Fix swapchain resize crash.
Forgot to update descriptorsets to point to the new recreated framebuffer attachments.

Currently ubo and material descriptors get updated on resize as well, which isn't necessarry, so they'll have to be moved out to separate functions. Could probably move the normal/albedo descriptors to the createrenderattachments function, move the ubo descriptor to the createuniformbuffer function, and move the model texture/material descriptors to the same place where we actually create the images.